+++ CHANGELOG.md
@@ -35,6 +35,7 @@ with the `YY.MINOR.MICRO` scheme.
   (See also https://git.picodata.io/core/picodata/-/issues/2273).
 - Fix `use-after-poison` in `proc_sql_execute`.
 - Fixed a bug which made the effective value of `pg_portal_max` be the same as `pg_statement_max`
+- Fix trigger execution and privilege initialization for scram-sha256 auth in pgproto.
 
 ### WebUI
 - Login form and simple session control